requirements.txt
requests
pyyaml
numpy
pandas
scikit-learn
sentence-transformers
tiktoken
